feat: small ws improvements (#59)
* feat: doc example and timestamp parsing

* feat: cors/logging layer + ping msg

* Update ohlc.rs

* feat: msg handling

* fix: clippy

* fix: taplo

+++ pragma-node/src/handlers/entries/get_entry.rs
@@ -34,7 +34,7 @@ pub async fn get_entry(
     // Construct pair id
     let pair_id = currency_pair_to_pair_id(&pair.0, &pair.1);
 
-    let now = chrono::Utc::now().timestamp() as u64;
+    let now = chrono::Utc::now().timestamp();
 
     let timestamp = if let Some(timestamp) = params.timestamp {
         timestamp
