+++ pragma-node/src/infra/repositories/entry_repository.rs
@@ -87,7 +87,7 @@ pub async fn routing(
     pool: &deadpool_diesel::postgres::Pool,
     pair_id: String,
     interval: Interval,
-    timestamp: u64,
+    timestamp: i64,
     is_routing: bool,
     agg_mode: AggregationMode,
 ) -> Result<(MedianEntry, u32), InfraError> {
@@ -164,7 +164,7 @@ async fn find_alternative_pair_price(
     base: &str,
     quote: &str,
     interval: Interval,
-    timestamp: u64,
+    timestamp: i64,
     agg_mode: AggregationMode,
 ) -> Result<(MedianEntry, u32), InfraError> {
     let conn = pool.get().await.map_err(adapt_infra_error)?;
@@ -213,7 +213,7 @@ async fn get_price_decimals(
     pool: &deadpool_diesel::postgres::Pool,
     pair_id: String,
     interval: Interval,
-    timestamp: u64,
+    timestamp: i64,
     agg_mode: AggregationMode,
 ) -> Result<(MedianEntry, u32), InfraError> {
     let entry = match agg_mode {
@@ -251,7 +251,7 @@ pub async fn get_twap_price(
     pool: &deadpool_diesel::postgres::Pool,
     pair_id: String,
     interval: Interval,
-    time: u64,
+    time: i64,
 ) -> Result<MedianEntry, InfraError> {
     let conn = pool.get().await.map_err(adapt_infra_error)?;
 
@@ -330,7 +330,7 @@ pub async fn get_twap_price(
         }
     };
 
-    let date_time = DateTime::from_timestamp(time as i64, 0).ok_or(InfraError::InvalidTimeStamp)?;
+    let date_time = DateTime::from_timestamp(time, 0).ok_or(InfraError::InvalidTimeStamp)?;
 
     let raw_entry = conn
         .interact(move |conn| {
@@ -357,7 +357,7 @@ pub async fn get_median_price(
     pool: &deadpool_diesel::postgres::Pool,
     pair_id: String,
     interval: Interval,
-    time: u64,
+    time: i64,
 ) -> Result<MedianEntry, InfraError> {
     let conn = pool.get().await.map_err(adapt_infra_error)?;
 
@@ -436,7 +436,7 @@ pub async fn get_median_price(
         }
     };
 
-    let date_time = DateTime::from_timestamp(time as i64, 0).ok_or(InfraError::InvalidTimeStamp)?;
+    let date_time = DateTime::from_timestamp(time, 0).ok_or(InfraError::InvalidTimeStamp)?;
 
     let raw_entry = conn
         .interact(move |conn| {
@@ -595,7 +595,7 @@ pub async fn get_ohlc(
     pool: &deadpool_diesel::postgres::Pool,
     pair_id: String,
     interval: Interval,
-    time: u64,
+    time: i64,
 ) -> Result<Vec<OHLCEntry>, InfraError> {
     let conn = pool.get().await.map_err(adapt_infra_error)?;
 
@@ -682,7 +682,7 @@ pub async fn get_ohlc(
         }
     };
 
-    let date_time = DateTime::from_timestamp(time as i64, 0).ok_or(InfraError::InvalidTimeStamp)?;
+    let date_time = DateTime::from_timestamp(time, 0).ok_or(InfraError::InvalidTimeStamp)?;
 
     let raw_entries = conn
         .interact(move |conn| {
